requirements.txt
# pyobjc
pynput
ruamel.yaml